+++ src/registry/domain/validators/index.ts
@@ -1,13 +1,12 @@
 import semver from 'semver';
 
 import componentParametersValidator from './component-parameters';
+import nodeVersionValidator from './node-version';
 import ocCliVersionValidator from './oc-cli-version';
 import packageJsonValidator from './package-json-validator';
 import pluginsRequirementsValidator from './plugins-requirements';
 import registryConfigurationValidator from './registry-configuration';
 import templateOcVersionValidator from './templace-oc-version';
-
-import nodeVersionValidator from './node-version';
 import uploadedPackageValidator from './uploaded-package';
 
 export function validateComponentName(componentName: string): boolean {